[move-lang] Update compiler API
- Overhaul compiler API for a 'Compiler' struct entry point
- Will make it a lot easier to deal with a extra flags

Closes: #8523
