+++ language/move-lang/tests/stdlib_sanity_check.rs
@@ -1,16 +1,13 @@
 // Copyright (c) The Diem Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use move_lang::{move_compile, shared::Flags};
+use move_lang::{Compiler, Flags};
 
 fn sanity_check_testsuite_impl(targets: Vec<String>, deps: Vec<String>) {
-    let (files, units_or_errors) = move_compile(
-        &targets,
-        &deps,
-        None,
-        Flags::empty().set_sources_shadow_deps(false),
-    )
-    .unwrap();
+    let (files, units_or_errors) = Compiler::new(&targets, &deps)
+        .set_flags(Flags::empty().set_sources_shadow_deps(false))
+        .build()
+        .unwrap();
     let errors = match units_or_errors {
         Err(errors) => errors,
         Ok(units) => move_lang::compiled_unit::verify_units(units).1,
