+++ language/tools/move-cli/src/base/commands/check.rs
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use anyhow::Result;
-use move_lang::{self, shared::Flags};
+use move_lang::{self, shared::Flags, Compiler};
 
 /// Type-check the user modules in `files` and the dependencies in `interface_files`
 pub fn check(
@@ -14,11 +14,8 @@ pub fn check(
     if verbose {
         println!("Checking Move files...");
     }
-    move_lang::move_check_and_report(
-        files,
-        interface_files,
-        None,
-        Flags::empty().set_sources_shadow_deps(sources_shadow_deps),
-    )?;
+    Compiler::new(files, interface_files)
+        .set_flags(Flags::empty().set_sources_shadow_deps(sources_shadow_deps))
+        .check_and_report()?;
     Ok(())
 }
