+++ language/tools/move-cli/src/base/commands/compile.rs
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use anyhow::Result;
-use move_lang::{self, shared::Flags};
+use move_lang::{self, shared::Flags, Compiler};
 
 /// Compile the user modules in `sources` against the dependencies in `interface_files`, placing
 /// the resulting binaries in `output_dir`.
@@ -17,11 +17,8 @@ pub fn compile(
     if verbose {
         println!("Compiling Move files...");
     }
-    let (files, compiled_units) = move_lang::move_compile_and_report(
-        sources,
-        interface_files,
-        None,
-        Flags::empty().set_sources_shadow_deps(sources_shadow_deps),
-    )?;
+    let (files, compiled_units) = Compiler::new(sources, interface_files)
+        .set_flags(Flags::empty().set_sources_shadow_deps(sources_shadow_deps))
+        .build_and_report()?;
     move_lang::output_compiled_units(emit_source_map, files, compiled_units, &output_dir)
 }
